Partial move of `cloneIfNecessaryModern()` into `fabric-value.ts`.
As of this commit, `cloneHelper()` is left as a dangling reference
in `-modern.ts`.
